requirements.txt
playwright-stealth
pika==1.3.2
playwright
openpyxl
httpx
requests
openpyxl
cos-python-sdk-v5
Pillow